Hooks for managing list/bookshelf display modes in playlists and collections
